convert browser.js to typescript (#44524)
* file renamed to .ts

* fixed ts errors in browser.ts
