+++ frontend/src/metabase/dashboard/components/Dashboard/Dashboard.tsx
@@ -97,7 +97,7 @@ type DashboardProps = {
   isAutoApplyFilters: boolean;
   isNavigatingBackToDashboard: boolean;
   addCardOnLoad?: DashCardId;
-  editingOnLoad?: string | string[];
+  editingOnLoad?: string | string[] | boolean;
 
   initialize: (opts?: { clearCache?: boolean }) => void;
   cancelFetchDashboardCardData: () => void;
